Created separate components for map list and map grid. I've also implemented vuex for library layout and renamed some files.
