+++ vue/src/store/mutation-types.js
@@ -6,3 +6,5 @@ export const ADD_MAP = 'ADD_MAP'
 export const REMOVE_MAP = 'REMOVE_MAP'
 export const REMOVE_ALL_MAPS = 'REMOVE_ALL_MAPS'
 export const UPDATE_MAP = 'UPDATE_MAP'
+
+export const SET_LAYOUT = 'SET_LAYOUT'
